+++ axelor-sale/src/main/java/com/axelor/apps/sale/db/repo/SaleOrderManagementRepository.java
@@ -69,11 +69,12 @@ public class SaleOrderManagementRepository extends SaleOrderRepository {
 	@Override
 	public SaleOrder save(SaleOrder saleOrder) {
 		try {
+		    saleOrder = super.save(saleOrder);
 			computeSeq(saleOrder);
 			computeFullName(saleOrder);
 			computeSubMargin(saleOrder);
 			Beans.get(SaleOrderMarginService.class).computeMarginSaleOrder(saleOrder);
-			return super.save(saleOrder);
+			return saleOrder;
 		} catch (Exception e) {
 			throw new PersistenceException(e.getLocalizedMessage());
 		}