+++ axelor-stock/src/main/java/com/axelor/apps/stock/service/InventoryService.java
@@ -353,7 +353,6 @@ public class InventoryService {
 					stockMoveLine.setTrackingNumber(trackingNumber);
 				}
 
-				stockMove.addStockMoveLineListItem(stockMoveLine);
 			}
 		}
 		if (stockMove.getStockMoveLineList() != null) {
@@ -382,8 +381,8 @@ public class InventoryService {
 
 	public StockMove createStockMoveHeader(Inventory inventory, Company company, StockLocation toStockLocation, LocalDate inventoryDate, String name) throws AxelorException  {
 
-		StockMove stockMove = Beans.get(StockMoveService.class).createStockMove(null, null, company, null,
-				stockConfigService.getInventoryVirtualStockLocation(stockConfigService.getStockConfig(company)), toStockLocation, inventoryDate, inventoryDate, null, null, null);
+		StockMove stockMove = Beans.get(StockMoveService.class).createStockMove(null, null, company,
+				stockConfigService.getInventoryVirtualStockLocation(stockConfigService.getStockConfig(company)), toStockLocation, inventoryDate, inventoryDate, null);
 
 		stockMove.setTypeSelect(StockMoveRepository.TYPE_INTERNAL);
 		stockMove.setName(name);
@@ -406,6 +405,15 @@ public class InventoryService {
 		if (stockLocationLineList != null) {
 			Boolean succeed = false;
 			for (StockLocationLine stockLocationLine : stockLocationLineList) {
+				if (stockLocationLine.getTrackingNumber() == null) { // if no tracking number on stockLocationLine, check if there is a tracking number on the product
+					long numberOfTrackingNumberOnAProduct = Beans.get(StockLocationLineRepository.class).all()
+							.filter("self.product = ?1 AND self.trackingNumber IS NOT null",
+									stockLocationLine.getProduct()).count();
+
+					if (numberOfTrackingNumberOnAProduct != 0) { // there is a tracking number on the product
+						continue;
+					}
+				}
 				inventory.addInventoryLineListItem(this.createInventoryLine(inventory, stockLocationLine));
 				succeed = true;
 			}