+++ axelor-supplychain/src/main/java/com/axelor/apps/supplychain/service/SaleOrderStockServiceImpl.java
@@ -124,7 +124,8 @@ public class SaleOrderStockServiceImpl implements SaleOrderStockService  {
 
 		StockMove stockMove = stockMoveService.createStockMove(null, saleOrder.getDeliveryAddress(), company,
 				saleOrder.getClientPartner(), saleOrder.getStockLocation(), toStockLocation, null, saleOrder.getShipmentDate(),
-				saleOrder.getDescription(), saleOrder.getShipmentMode(), saleOrder.getFreightCarrierMode());
+				saleOrder.getDescription(), saleOrder.getShipmentMode(), saleOrder.getFreightCarrierMode(), 
+				saleOrder.getCarrierPartner(), saleOrder.getForwarderPartner(), saleOrder.getIncoterm());
 
 		stockMove.setToAddressStr(saleOrder.getDeliveryAddressStr());
 		stockMove.setSaleOrder(saleOrder);
@@ -217,7 +218,6 @@ public class SaleOrderStockServiceImpl implements SaleOrderStockService  {
 			if (stockMoveLine != null) {
 	            stockMoveLine.setSaleOrderLine(saleOrderLine);
 	            stockMoveLine.setReservedQty(saleOrderLine.getReservedQty());
-	            stockMove.addStockMoveLineListItem(stockMoveLine);
 			}
 
 			return stockMoveLine;
@@ -236,9 +236,6 @@ public class SaleOrderStockServiceImpl implements SaleOrderStockService  {
 			saleOrderLine.setDeliveryState(SaleOrderRepository.STATE_NOT_DELIVERED);
 			stockMoveLine.setSaleOrderLine(saleOrderLine);
 
-			if(stockMoveLine != null) {
-				stockMove.addStockMoveLineListItem(stockMoveLine);
-			}
 			return stockMoveLine;
 		}
 		return null;